Added support for Debutant entry.
You can now mark a Driver as "Debutant" which will exclude the Driver from the regular points calculation and instead just get points by finishing the event. Amount of points given to Debutants is set in InitialState.js with "numPointsForDebutant: 5,".

You mark a Driver as Debutant by setting isDebutant: true, as an entry for the specific driver and event in manualResults